Finished Zone-based Movement
The WorldMap consists of dynamically spawning Zones that are populated with randomly spawning room sets drawn from a zone type fed to the Zone. There is currently only one zone type.

Each Zone Type Room Set needs 9 rooms to be complete.

As a player explores beyond the edges of a Zone, they will spawn a new zone, which will populate rooms in a random order. The doors of these rooms will line up with the doors of adjacent rooms and zones. 

The world map will dynamically expand as the player reaches its edge.