Full-text input actions and began add/remove inventory process
Started the process of full-text parsing for action input. Currently the function finds the first verb and noun and acts from there. This will need to be changed to include regex patterns to search for inventory items in addition to nouns (so that "inspect <ITEM>" or "equip <ITEM>" can also be used).

Added add/remove to inventory.